requirements.txt
ansi2html==1.9.1
ansible==8.1.0
ansible-base==2.10.17
ansible-core==2.15.1
ansible-pylibssh==1.1.0
ansible-runner==2.3.4
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-compat==1.0.15
django-environ==0.10.0
djangorestframework==3.14.0
docutils==0.20.1
httpie==3.2.2
idna==3.4
Jinja2==3.1.2
jxmlease==1.0.3
lockfile==0.12.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
ncclient==0.6.13
oauthlib==3.2.2
packaging==23.1
paramiko==3.2.0
pexpect==4.9.0
Pillow==10.0.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
PyNaCl==1.5.0
PySocks==1.7.1
python-daemon==3.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
resolvelib==1.0.1
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
xmltodict==0.13.0
